requirements.txt
astroid==2.12.13
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.0.12
dill==0.3.6
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
idna==3.4
importlib-metadata==5.0.0
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.8.0
mccabe==0.7.0
packaging==21.3
Pillow==8.3.1
platformdirs==2.5.4
pluggy==0.13.1
py==1.11.0
PyJWT==2.1.0
pylint==2.15.6
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2022.6
requests==2.26.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
typed-ast==1.5.4
typing_extensions==4.4.0
urllib3==1.26.12
wrapt==1.14.1
zipp==3.10.0
